+++ src/plonemeeting/restapi/tests/test_services_meetingconfig.py
@@ -38,14 +38,14 @@ class testServiceConfig(BaseTestCase):
     def test_restapi_config_endpoint(self):
         """@config"""
         cfg = self.meetingConfig
-        cfgId = cfg.getId()
-        endpoint_url = "{0}/@config?config_id={1}".format(self.portal_url, cfgId)
+        cfg_id = cfg.getId()
+        endpoint_url = "{0}/@config?config_id={1}".format(self.portal_url, cfg_id)
         response = self.api_session.get(endpoint_url)
         self.assertEqual(response.status_code, 200, response.content)
         json = response.json()
         # by default, no "items"
         self.assertFalse("items" in json)
-        self.assertEqual(json["id"], cfgId)
+        self.assertEqual(json["id"], cfg_id)
         self.assertEqual(json["UID"], cfg.UID())
 
     def test_restapi_config_extra_include_categories(self):
@@ -53,11 +53,11 @@ class testServiceConfig(BaseTestCase):
         # cfg2 uses disabled categories
         cfg = self.meetingConfig2
         cfg.setUseGroupsAsCategories(False)
-        cfgId = cfg.getId()
+        cfg_id = cfg.getId()
         transaction.commit()
         # categories
         endpoint_url_pattern = "{0}/@config?config_id={1}&extra_include={2}"
-        endpoint_url = endpoint_url_pattern.format(self.portal_url, cfgId, "categories")
+        endpoint_url = endpoint_url_pattern.format(self.portal_url, cfg_id, "categories")
         response = self.api_session.get(endpoint_url)
         self.assertEqual(response.status_code, 200, response.content)
         json = response.json()
@@ -72,13 +72,31 @@ class testServiceConfig(BaseTestCase):
         self.assertEqual(categories[-2]["enabled"], False)
         self.assertEqual(categories[-2]["review_state"], None)
 
+    def test_restapi_config_extra_include_classifiers(self):
+        """@config"""
+        cfg = self.meetingConfig
+        cfg_id = cfg.getId()
+        endpoint_url_pattern = "{0}/@config?config_id={1}&extra_include={2}"
+        endpoint_url = endpoint_url_pattern.format(self.portal_url, cfg_id, "classifiers")
+        response = self.api_session.get(endpoint_url)
+        self.assertEqual(response.status_code, 200, response.content)
+        json = response.json()
+        classifiers = json["extra_include_classifiers"]
+        self.assertEqual(len(classifiers), 3)
+        self.assertEqual(classifiers[0]["id"], u"classifier1")
+        self.assertEqual(classifiers[0]["@type"], u"meetingcategory")
+        self.assertEqual(classifiers[0]["enabled"], True)
+        self.assertEqual(classifiers[0]["review_state"], None)
+        self.assertEqual(classifiers[1]["id"], u"classifier2")
+        self.assertEqual(classifiers[2]["id"], u"classifier3")
+
     def test_restapi_config_extra_include_pod_templates(self):
         """@config"""
         cfg = self.meetingConfig
-        cfgId = cfg.getId()
+        cfg_id = cfg.getId()
         endpoint_url_pattern = "{0}/@config?config_id={1}&extra_include={2}"
         endpoint_url = endpoint_url_pattern.format(
-            self.portal_url, cfgId, "pod_templates"
+            self.portal_url, cfg_id, "pod_templates"
         )
         response = self.api_session.get(endpoint_url)
         self.assertEqual(response.status_code, 200, response.content)
@@ -106,9 +124,9 @@ class testServiceConfig(BaseTestCase):
     def test_restapi_config_extra_include_searches(self):
         """@config extra_include=searches"""
         cfg = self.meetingConfig
-        cfgId = cfg.getId()
+        cfg_id = cfg.getId()
         endpoint_url_pattern = "{0}/@config?config_id={1}&extra_include={2}"
-        endpoint_url = endpoint_url_pattern.format(self.portal_url, cfgId, "searches")
+        endpoint_url = endpoint_url_pattern.format(self.portal_url, cfg_id, "searches")
         response = self.api_session.get(endpoint_url)
         self.assertEqual(response.status_code, 200, response.content)
         json = response.json()
@@ -136,9 +154,9 @@ class testServiceConfig(BaseTestCase):
     def test_restapi_config_extra_include_proposing_groups(self):
         """@config extra_include=proposing_groups"""
         cfg = self.meetingConfig
-        cfgId = cfg.getId()
+        cfg_id = cfg.getId()
         endpoint_url_pattern = "{0}/@config?config_id={1}&extra_include={2}"
-        endpoint_url = endpoint_url_pattern.format(self.portal_url, cfgId, "proposing_groups")
+        endpoint_url = endpoint_url_pattern.format(self.portal_url, cfg_id, "proposing_groups")
         response = self.api_session.get(endpoint_url)
         self.assertEqual(response.status_code, 200, response.content)
         json = response.json()
@@ -149,9 +167,9 @@ class testServiceConfig(BaseTestCase):
     def test_restapi_config_extra_include_associated_groups(self):
         """@config extra_include=associated_groups"""
         cfg = self.meetingConfig
-        cfgId = cfg.getId()
+        cfg_id = cfg.getId()
         endpoint_url_pattern = "{0}/@config?config_id={1}&extra_include={2}"
-        endpoint_url = endpoint_url_pattern.format(self.portal_url, cfgId, "associated_groups")
+        endpoint_url = endpoint_url_pattern.format(self.portal_url, cfg_id, "associated_groups")
         response = self.api_session.get(endpoint_url)
         self.assertEqual(response.status_code, 200, response.content)
         json = response.json()
@@ -162,11 +180,11 @@ class testServiceConfig(BaseTestCase):
     def test_restapi_config_extra_include_groups_in_charge(self):
         """@config extra_include=groups_in_charge"""
         cfg = self.meetingConfig
-        cfgId = cfg.getId()
+        cfg_id = cfg.getId()
         self._enableField('groupsInCharge')
         transaction.commit()
         endpoint_url_pattern = "{0}/@config?config_id={1}&extra_include={2}"
-        endpoint_url = endpoint_url_pattern.format(self.portal_url, cfgId, "groups_in_charge")
+        endpoint_url = endpoint_url_pattern.format(self.portal_url, cfg_id, "groups_in_charge")
         response = self.api_session.get(endpoint_url)
         self.assertEqual(response.status_code, 200, response.content)
         json = response.json()
@@ -177,16 +195,16 @@ class testServiceConfig(BaseTestCase):
     def test_restapi_config_extra_include_fullobjects(self):
         """@config extra_include_fullobjects"""
         cfg = self.meetingConfig
-        cfgId = cfg.getId()
+        cfg_id = cfg.getId()
         self._enableField('groupsInCharge')
         cfg.setUseGroupsAsCategories(False)
-        cfgId = cfg.getId()
+        cfg_id = cfg.getId()
         transaction.commit()
         endpoint_url_pattern = "{0}/@config?config_id={1}&extra_include={2}" \
             "&extra_include={3}&extra_include={4}&extra_include={5}&extra_include={6}"
         endpoint_url = endpoint_url_pattern.format(
             self.portal_url,
-            cfgId,
+            cfg_id,
             "categories",
             "pod_templates",
             "searches",